test: split the suite by where the logic lives, not by layer
Every module carried a unit test per layer — repository, service, controller —
regardless of whether it had anything of its own to assert. For a plain CRUD
module those tests exercised the shared machinery through a module that adds
nothing to it: they broke whenever the framework changed shape, and caught no
bugs doing it.

Two levels now, and the line between them is where the logic lives. Unit covers
the shared machinery and a module's own business rules. E2e covers each
entity's flow over HTTP, which is what proves the wiring holds — route mounted,
guard applied, validation running, status codes and error envelope — without
enumerating cases.

- The flow suites moved from tests/integration to tests/e2e, which is what they
  always were, and e2e is enabled in jest instead of commented out. The old
  32-line e2e file went: it asserted /api/users answered 200 with no token.
- Three per-entity suites are gone: users' repository test and the branches and
  appointments controller tests, 365 lines asserting delegation, id parsing and
  404s that CrudService, CrudController and BaseModuleRepository already assert.
- The rules stay where the cases are. Appointments keeps its 448 lines of
  overlap edge cases and branches its schedule and cascade rules; reproducing
  those over HTTP would be slower, longer and more fragile.

Coverage went from 95.2% to 94.6% of statements for 365 fewer lines of test,
because the e2e was already covering the same code. The per-layer thresholds
still pass.

docs/testing.md and the add-a-module guide now state the rule, including why
there is no controller unit test in the list.
